Move cv2_to_ros and ros_to_cv2 methods to utilities.
