+++ ros2_ws/src/rcdt_detection/CMakeLists.txt
@@ -9,6 +9,7 @@ project(rcdt_detection)
 find_package(rclpy REQUIRED)
 find_package(ament_cmake REQUIRED)
 find_package(ament_cmake_python REQUIRED)
+find_package(rcdt_utilities REQUIRED)
 
 # Install project files
 install(DIRECTORY
